+++ src/ast/collect_annotation_tyvars.rs
@@ -1,5 +1,4 @@
 use std::sync::Arc;
-
 use crate::{
     ast::{
         expr::{Expr, ExprNode},