+++ src/optimization/remove_hktvs.rs
@@ -21,7 +21,6 @@ Purpose:
 */
 
 use std::sync::Arc;
-
 use crate::{
     ast::{
         expr::{expr_let_typed, expr_match_typed, Expr, ExprNode},